create controller and fix small code errors
